+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Fixed migration sales script for separate payment methods.
 * Fixed migration sales script for Svea service payment selection method.
 * Only offline refunds available -text removed from the admin order page.
+* Show message on admin page if Svea Seller ID is not set.
+* Clear cached payment methods if Svea Payments settings change.
 
 ### [1.0.15] - 2024-11-14
 * Enable changing amount of handling fee and handling fee tax to refund.